requirements.txt
asttokens==2.4.1
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.13.0
beautifulsoup4==4.12.3
boto3==1.34.16
botocore==1.34.60
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
loguru==0.7.2
matplotlib-inline==0.1.6
msal==1.27.0
msal-extensions==1.1.0
msrest==0.7.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.3
pdfminer.six==20231228
pexpect==4.9.0
platformdirs==4.2.0
playwright==1.43.0
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.10.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
XlsxWriter==3.2.0
